Move publishing code into contentcuration.utils and make a publish root function in preparation for progress tracking integration.
